+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Show the report button on the own profile page !2589 @alex.simm
 - Hacky workaround to replace $amp in urls from chat messages #1531 !2594 @chriswalg
 - Fixed some stability issues with incoming Mail processing !2598 !2607 @\_fridtjof_
+- Removed remains of mapbox-gl in the LeafletMap component #1570 !2628 @alex.simm
 
 
 ## Refactoring